add bubble sort and refactor binary search into its own component
